+++ src/test/C/Laplace/templaplace.cpp
@@ -12,16 +12,16 @@ void LaplaceFor(
     {
       for (unsigned j = 0; j < storagesize; j++)
         {
-          float gradient_temp[dim];
-          float dot_temp[dim];
+          double gradient_temp[dim];
+          double dot_temp[dim];
           for (unsigned d = 0; d < dim; d++)
             {
-              float level_i = level[i][d];
-              float level_j = level[j][d];
-              float level_int_i = level_int[i][d];
-              float level_int_j = level_int[j][d];
-              float index_i = index[i][d];
-              float index_j = index[j][d];
+              double level_i = level[i][d];
+              double level_j = level[j][d];
+              double level_int_i = level_int[i][d];
+              double level_int_j = level_int[j][d];
+              double index_i = index[i][d];
+              double index_j = index[j][d];
               gradient_temp[d] = gradient(
 	level_i, index_i, level_j, 
 	index_j, lcl_q_inv[d]);
@@ -30,10 +30,10 @@ void LaplaceFor(
 	index_j, level_int_i, level_int_j, 
 	lcl_q[d]);
             }
-          float sub = 0.0;
+          double sub = 0.0;
           for (size_t d_outer = 0; d_outer < dim; d_outer++)
             {
-              float element = alpha[j];
+              double element = alpha[j];
               for (size_t d_inner = 0; d_inner < dim; d_inner++)
                 {
                   element *= (dot_temp[d_inner] * (d_outer != d_inner)) + (gradient_temp[d_inner] * (d_outer == d_inner));
